Adds property and user services
Implements services for managing properties and users.

The PropertyService provides functionalities to create, retrieve, update, and delete property listings, as well as to retrieve them based on owner or city.
It also includes logic for associating properties with users, publications, addresses, photos and amenities.

The UserService provides functionalities to create, retrieve, update, and delete users.